requirements.txt
flask
flask_restful
flask_sqlalchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
